Micser_Python/requirements.txt
grpcio>=1.76.0
grpcio-tools>=1.62.1
protobuf>=6.31.1
openai>=1.0.0
requests>=2.28.0
sqlalchemy>=1.4.0
asyncpg>=0.27.0
pydantic>=1.10.0
python-dotenv>=1.0.0
perplexityai==0.20.0